refactor: extract SidebarSelection and ServerStatusView
Split SourceListView.swift (271 lines) by moving the SidebarSelection
enum and the ServerStatusView struct into their own files for better
discoverability and single-responsibility. Same module, no import
changes needed elsewhere. Also removed an unused ProcessManager
environment binding from SourceListView left behind by the split.
