+++ apps/mac-client/SuperPickyApp/SidebarSelection.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum SidebarSelection: Hashable {
+    case folder(URL)
+    case rating(Int)
+    case flying
+    case picks
+    case species(String)
+    case burstGroup(UUID)
+}
