+++ apps/mac-client/SuperPickyApp/SourceListView.swift
@@ -1,14 +1,5 @@
 import SwiftUI
 
-enum SidebarSelection: Hashable {
-    case folder(URL)
-    case rating(Int)
-    case flying
-    case picks
-    case species(String)
-    case burstGroup(UUID)
-}
-
 struct SourceListView: View {
     @Binding var selection: SidebarSelection?
     @Binding var folders: [URL]
@@ -18,7 +9,6 @@ struct SourceListView: View {
     let speciesEntries: [SpeciesEntry]
     let processingFolder: URL?
     let processingProgress: Double
-    @Environment(ProcessManager.self) private var processManager
 
     let onAddFolder: () -> Void
     let onRemoveFolder: (URL) -> Void
@@ -241,31 +231,3 @@ struct FolderRow: View {
     }
 }
 
-struct ServerStatusView: View {
-    @Environment(ProcessManager.self) private var processManager
-
-    var body: some View {
-        HStack(spacing: 6) {
-            Circle()
-                .fill(processManager.isReady ? .green : (processManager.isRunning ? .orange : .red))
-                .frame(width: 8, height: 8)
-
-            Text(statusText)
-                .font(.caption2)
-                .foregroundStyle(.secondary)
-
-            Spacer()
-        }
-        .accessibilityIdentifier("ServerStatus")
-    }
-
-    private var statusText: String {
-        if processManager.isReady {
-            return "Models ready"
-        } else if processManager.isRunning {
-            return "Loading models..."
-        } else {
-            return "Server offline"
-        }
-    }
-}
